Dp sf pap (#544)
* use both _1 and _2 segmentations

* use both _1 and _2 segmentations

* TEMP: add dice metrics, copied from neuron

* ENH: Use dice loss and metrics

* ENH: Remove kidney label and merge body/background labels

* FIX: Fix bad number of channels

* ENH: Use only one channel of the input image

* WIP: hacking bottom of U-net

* ENH: Add mean and std for normalization

* ENH: Add neurite and voxelmorph to docker

* FIX: Use dice loss from neurite

* STYLE: Fix up WIP code on hacking bottom of U-net

* ENH: Add merged paps for segmentation tensormap

* WIP: Fix Unet concats

* FIX: Fix soft dice metrics

* ENH: Add plot_dice to compare

* ENH: Add median computation for papillary segmentation project

* FIX: Fix double plot on one graph

* ENH: Allow generator to have empty path, e.g., to test on all images

* ENH: Prune list of structures for which we do stats

* STYLE: rearranging

* WIP: Handle inference without ground truth labels

* ENH: Remove option for merged paps

* FIX: Get all b2s images, instance_2s only

* ENH: Add mri dates

* FIX: Fix normalization with correct padding

* FIX: Fix soft dice metrics again

* COMP: Add option for environment variable for jupyter notebooks

* WIP: data augmentation

* WIP: Better scatter plots for medians

* ENH: Report std too

* ENH: Improve dice plots for a single model

* ENH: Log pearson correlation coefficients

* STYLE: Adding TODOs to fix tensor_generators

* WIP: Add temporary code to save Dice scores

* WIP: Add temporary code for plotting medians

* STYLE: Clean up code for infer_medians

* STYLE: Clean up medians code

* STYLE: Add command-line args for median computations

* ENH: Add percentiles and tsv for dice calculations

* STYLE: Add command-line args for data augmentation

* ENH: Improve log files for dice compare

* STYLE: Small edits requested in PR

* STYLE: docstring and typehints for plot_dice

* STYLE: docstring for infer_statistics_from_segmented_regions

* STYLE: Docstring and typehints for augment_using_layers

---------

+++ ml4h/tensorize/tensor_writer_ukbb.py
@@ -664,7 +664,7 @@ def _get_overlay_from_dicom(d, debug=False) -> Tuple[np.ndarray, np.ndarray]:
 
 def _unit_disk(r) -> np.ndarray:
     y, x = np.ogrid[-r: r + 1, -r: r + 1]
-    return (x ** 2 + y ** 2 <= r ** 2).astype(np.int)
+    return (x ** 2 + y ** 2 <= r ** 2).astype(np.int32)
 
 
 def _outline_to_mask(labeled_outline, idx) -> np.ndarray:
